Bumped typarse to 3.4
Centralized all the BaseConfig's in a single file to avoid repeating the same values in defaults
