+++ coltra/TODO.txt
@@ -1,10 +1,11 @@
-TODO: wandb integration (finish - some videos)
 
 TODO: Fix the configs (1: model arguments, 2: more sane structure to avoid having to manually propagate the changes in typarsers)
+Big controversial TODO: remove all default values from configs
 
+TODO: optuna integration
 TODO: Implement ugae  that works for arbitrary episode lengths
 
-Big controversial TODO: remove all default values from configs
+TODO: wandb integration (finish - some videos)
 
 TODO: Integrate Action/Observation abstractions with gym.spaces
