+++ scripts/configs/base_config.yaml
@@ -38,7 +38,7 @@ trainer:
 
 model:
   input_size:
-  num_actions: 2
+  num_actions:
   activation: tanh
   discrete: false
