+++ scripts/train_gym.py
@@ -100,5 +100,5 @@ if __name__ == "__main__":
     if CUDA:
         agent.cuda()
 
-    trainer = PPOCrowdTrainer(agent, env, config)
+    trainer = PPOCrowdTrainer(agent, env, trainer_config)
     trainer.train(args.iters, disable_tqdm=False, save_path=trainer.path)
