+++ tests/test_models.py
@@ -7,7 +7,7 @@ from typarse import BaseConfig
 
 from coltra.agents import CAgent
 from coltra.buffers import Observation
-from coltra.models import FCNetwork
+from coltra.models import FCNetwork, FancyMLPModel
 from coltra.models.raycast_models import LeeNetwork, LeeModel
 from coltra.models.relational_models import RelationNetwork, RelationModel
 
@@ -103,3 +103,19 @@ def test_relnet():
     assert action.scale.shape == (7, 2)
     assert state == ()
     assert extra["value"].shape == (7, 1)
+
+
+def test_multiple_mlps():
+    config1 = {"input_size": 3, "num_actions": 1, "discrete": False}
+
+    mlp1 = FancyMLPModel(config1)
+
+    assert mlp1.discrete is False
+    assert mlp1.policy_network.hidden_layers[0].in_features == 3
+
+    config2 = {"input_size": 5, "num_actions": 2, "discrete": True}
+
+    mlp2 = FancyMLPModel(config2)
+
+    assert mlp2.discrete is True
+    assert mlp2.policy_network.hidden_layers[0].in_features == 5
